Simple scripts to load and retrieve data from a store.
These are used to get data into a store for testing, and
then profile the simple retrieval of all the tiddlers in
a bag.

These confirm the expected result, which of the tiddlyweb code,
most time is spent in `tiddler_get` of the store. Note this
is only profiling the code reached by getsome, which is not
representative of an active server, just one specific thing.

This isn't news but was a useful datapoint.